index: Implement file index load
This completes the implementation of single-file index handling.
Loads the index back into ram, and enhances find() and the
SortedIterator to be able to search both the data loaded from the file
and the data added since in ram.

Tests the various combinations of ram and file data with all searching
for both failing and succeeding cases, which should hit all of the
edge conditions on the binary search.  The only case missing is when
the file is very small and many of the top index values are zero.
This would be tested by not inserting much, and testing read with
radically different keys.