Implement pool file locking
This uses the underlying lockf() C call, and requires some libstdc++
trickery to get to the file descriptor used for a file.  This should
be fairly safe, and generate a compile error if the implementation
details change to where it wouldn't work.

It seems to work correctly with clang, and to segfault with gcc.