pool: Fix UUID generator's random source
The default random UUID generator in boost seems to be based on using
ininitialized data in the Mersenne Twister.  Aside from being
unreliable, this also throws a bunch of warnings in valgrind.

Fix by using the 'random_device' generator, which gets the random data
from /dev/urandom on Linux.  This should be a much better source for
random uuids.