chunk: Simulate const accessors
Several methods of Chunk perform an expensive
compression/decompression operation, and therefore modify the internal
state of the Chunk.  However, this operation is idempotent, and the
class behaves as if the operations were constant.

This is similar to the way values are updated in lanauges with lazy
evaluation.  The values themselves are immutable, but the results are
cached.

Implement this, by declaring the accessors as const methods.  When
updating the cache internally, use a const_cast to poke under the hood
and update the cached values.